debug

+++ train_llama2_multiHead_loop.py
@@ -318,10 +318,11 @@ class LlamaForSequenceClassificationMultiHead(LlamaPreTrainedModel):
         else:
             if input_ids is not None:
                 # find the sequence length by finding the first end padding. Note that pad shares the same token with eos.
+                # the idea is that find the first `<pad>` in the subsequent `<pad>, <pad>, ..., <pad>`,
                 is_pad_token = torch.eq(input_ids, self.config.pad_token_id).long()
                 diff = is_pad_token[:, :-1] - is_pad_token[:, 1:] * 2
                 print(f"{diff=}")
-                target_id = (diff == -1).argmin(-1).to(logits1.device)
+                target_id = (diff == -1).long().argmin(-1).to(logits1.device)
                 sequence_lengths = (torch.eq(input_ids, self.config.pad_token_id).long().argmax(-1) - 1).to(
                     logits1.device
                 )
